// src/models/analytics.ts
export interface AnalyticsSample {
  dayUnrlzd: number;
  dayRlzd: number;
  dayInt: number;
  dayFx: number;
  dayPNL: number;

  mtdUnrlzd: number;
  mtdRlzd: number;
  mtdInt: number;
  mtdFx: number;
  mtdPNL: number;

  dayUnrlzdOfNAV: number;
  dayRlzdOfNAV: number;
  dayIntOfNAV: number;
  dayFxOfNAV: number;
  dayPNLOfNAV: number;

  mtdUnrlzdOfNAV: number;
  mtdRlzdOfNAV: number;
  mtdIntOfNAV: number;
  mtdFxOfNAV: number;
  mtdPNLOfNAV: number;

  dayUnrlzdUniform: number;
  dayRlzdUniform: number;
  dayIntUniform: number;
  dayFxUniform: number;
  dayPNLUniform: number;

  mtdUnrlzdUniform: number;
  mtdRlzdUniform: number;
  mtdIntUniform: number;
  mtdFxUniform: number;
  mtdPNLUniform: number;
}